LayoutTests/platform/chromium/http/tests/security/dataURL/xss-DENIED-from-data-url-in-foreign-domain-window-open-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://127.0.0.1:8000/security/dataURL/xss-DENIED-from-data-url-in-foreign-domain-window-open.html from frame with URL data:text/html,<html><head><script>window.onload = function(){try {parent.opener.document.getElementById('accessMe').innerHTML = 'FAIL: Cross frame access from a data: URL on a different domain was allowed';alert('FAIL: No exception thrown.');} catch (e) {alert('PASS: Exception thrown successfully.');}if (window.testRunner)testRunner.globalFlag = true;}</script></head><body><p>Inner-inner iframe. This iframe (which is data: URL and whose parent is on a foreign domain) is the frame attempting to access the main frame.  It should not have access to it.</p></body></html>. The frame requesting access has a protocol of '', the frame being accessed has a protocol of 'http'. Protocols must match.

ALERT: PASS: Exception thrown successfully.
Opener Frame

Pass: Cross frame access from a data: URL on a different domain was denied.
